soc/intel/cmn/acpi: exclude TPM area range if CRB TPM driver is enabled
The CRB TPM driver already adds its memory range to the SSDT. Under Windows,
this leads to a conflicting address range error, preventing the TPM from
being initialized.

In case the CRB TPM driver is enabled and the TPM base address is set to
0xfed40000, do not include this range in the northbridge ASL. Otherwise,
the behavior remains unchanged.

TEST=Build CB with CONFIG_CRB_TPM=y, CONFIG_CRB_TPM_BASE_ADDRESS=0xfed40000
and boot Windows 11, verify that the TPM doesn't show any errors in the
Device Manager.

Signed-off-by: Filip Brozovic <fbrozovic@gmail.com>