soc/intel/common/block/smbus: Keep TCO WDT timeout flag if WDAT_WDT=y
The TCO SMI handler clears the watchdog timeout flag unconditionally. Since the system
is only rebooted if the flag is already set and the watchdog timer expires again, this
means that the reboot never occurs. This change preserves the timeout flag if
CONFIG_WDAT_WDT is enabled, otherwise the behavior remains unchanged.

Signed-off-by: Filip Brozovic <fbrozovic@gmail.com>